Cleaned up repo folder structure
Deleted the 2 src folders inside the IntroToOpenGL project and
CameraAndProjection Projects. Linked all those projects to the same
folder that the RenderingGeometry Project is linked to and thats the src
folder inside the graphics folder. There was no reason to have copies of
the same files in 3 places and each project linking to a different copy.
So i deleted 2 src folder copies and kept 1. I just linked all 3
projects to the same one. Also deleted my git ignore because the
teachers was way better and useful